requirements.txt
requests>=2.31.0
mutagen>=1.47.0
tqdm>=4.66.1
PyQt6>=6.6.1
python-dotenv>=1.0.0 